pyproject.toml
[project]
name = "zara"
version = "0.1.0"
description = "Zara JP discount monitor"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "apscheduler>=3.11,<4",
    "python-dotenv>=1.1",
    "loguru>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zara"]

[project.scripts]
zara = "zara.cli:main"
